Reinstate `height: 100%` on the `<body>`
Reinstates the property and value which were removed as part of
https://github.com/fac/origin/releases/tag/1.1.4

Through further usage, it’s become obvious that having this rule is
important, as without it any `background-color` on the `<body>` won’t
span the full height of the `<html>` element.

Note that this rule can be overridden per-application. This change
won’t break anything., but this is the behaviour we want out of the box.